BIG FIX
updated kit
fixed issue with mod not showing in zues
updated the addons to active curator
