+++ @AH-64D Apache Longbow/Addons/fza_ah64_controls/config.cpp
@@ -3,11 +3,22 @@ class CfgPatches
 {
 	class fza_ah64_controls
 	{
-		units[] = {"fza_ah64base","fza_ah64d_b2e","fza_ah64d_b2e_nr"};
+		units[] = {"fza_ah64base","fza_ah64d_b2e","fza_ah64d_b2e_nr","fza_ah64_pilot"};
 		author = "Franze, Nodunit, Sacha 'Voodooflies' Oropeza, Keplager & Community";
 		weapons[] = {};
 		requiredVersion = 1.0;
-		requiredAddons[] = {"A3_Air_F_Beta","A3_Sounds_F","A3_Data_F", "cba_main", "cba_xeh", "fza_ah64_sfmplus"};
+		requiredAddons[] = {"A3_Air_F_Beta","A3_Sounds_F","A3_Data_F", "cba_main", "cba_xeh", "fza_ah64_sfmplus", "fza_ah64_AICrew"};
+	};
+};
+class CfgAddons
+{
+	class PreloadBanks{};
+	class PreloadAddons
+	{
+		class fza_ah64_controls
+		{
+			list[] = {"fza_ah64_controls", "fza_ah64_sfmplus", "fza_ah64_AICrew"};
+		};
 	};
 };
